refactoring stuff
