+++ src/components/navbar/Navbar.css
@@ -14,10 +14,14 @@
 }
 
 .navbar-brand {
-  font-size: 1.5rem;
-  font-weight: 700;
-  color: #006acc;
+  display: flex;
+  align-items: center;
   text-decoration: none;
+}
+
+.navbar-logo {
+  height: 40px;
+  width: auto;
 }
 
 .sign-in-button {
@@ -33,4 +37,4 @@
 
 .sign-in-button:hover {
   background-color: #0055b3;
-} 
+}
